feat: add shared content package
Move the admin content inventory and newsletter draft data into @anipotts/content so content/editor schema stops living directly inside the admin app. Keep app-level compatibility re-exports for current Astro routes.

Also teach automerge and deploy target calculation that packages/content changes affect the admin target.

Verified: git diff --check; pnpm exec prettier --check changed files; pnpm turbo typecheck --filter=@anipotts/content...; pnpm turbo build --filter=@anipotts/content...; pnpm turbo typecheck --filter=@anipotts/admin...; pnpm turbo build --filter=@anipotts/admin....
