+++ apps/admin/package.json
@@ -12,6 +12,7 @@
     "typecheck": "astro check"
   },
   "dependencies": {
+    "@anipotts/content": "workspace:*",
     "@astrojs/cloudflare": "^12.6.0",
     "@simplewebauthn/browser": "13.3.0",
     "@simplewebauthn/server": "13.3.2",
