+++ apps/admin/src/data/newsletter.ts
@@ -1,173 +1,4 @@
-export type NewsletterClaim = {
-  claim: string;
-  status: "source_backed" | "needs_proof";
-  source_refs: string[];
-};
-
-export type NewsletterSection = {
-  kind: "opening" | "note" | "receipt" | "worklog" | "close";
-  heading?: string;
-  label?: string;
-  body: string;
-  items?: string[];
-  source_refs?: string[];
-  thread_beat?: string;
-};
-
-export type NewsletterDraft = {
-  id: string;
-  slug: string;
-  status: "idea" | "draft" | "preview" | "ready_for_review" | "blocked";
-  title: string;
-  subject: string;
-  summary: string;
-  dek: string;
-  audience: string;
-  source_refs: { ref: string; use: string }[];
-  spine: string[];
-  sections: NewsletterSection[];
-  claims: NewsletterClaim[];
-  blocked_actions: string[];
-  preview_notes: string;
-  source_fixture: string;
-  preview_fixture: string;
-};
-
-export const newsletterDraftsSource = {
-  source_doc: "docs/newsletter-content-structure.md",
-  system_doc: "docs/newsletter-system.md",
-  mode: "read_only_static_draft_preview",
-  blocked:
-    "No D1 write, archive publish, send, schedule, worker, endpoint, DNS, auth, env, or secret mutation.",
-};
-
-export const newsletterDrafts: NewsletterDraft[] = [
-  {
-    id: "newsletter-draft-first-thing-agents-need-control-plane",
-    slug: "first-thing-agents-need-control-plane",
-    status: "idea",
-    title: "the first thing your agents need is a control plane",
-    subject: "the first thing your agents need is a control plane",
-    summary:
-      "every coding agent gets more useful when it knows what it is allowed to do, what proof it owes, and when it needs to come back to me.",
-    dek: "what i am learning from turning admin.anipotts.com into a read-only operator dashboard first, then a content review surface, then a guarded write path.",
-    audience:
-      "builders running coding agents across real repos, live sites, and operational state",
-    source_fixture:
-      "docs/newsletter-drafts/first-thing-agents-need-control-plane.json",
-    preview_fixture:
-      "docs/newsletter-drafts/first-thing-agents-need-control-plane.preview.html",
-    source_refs: [
-      {
-        ref: "AGENTS.md",
-        use: "repo authority model, safe lanes, hard gates, NEEDS-ANI rules, and proof expectations",
-      },
-      {
-        ref: "docs/admin-v2-architecture.md",
-        use: "admin v2 direction and control-plane framing",
-      },
-      {
-        ref: "docs/content-admin-editor-brief.md",
-        use: "read-only content inventory and admin editing progression",
-      },
-      {
-        ref: "apps/admin/src/data/content.ts",
-        use: "Astro admin static content inventory and preview-only model",
-      },
-      {
-        ref: "docs/newsletter-content-structure.md",
-        use: "newsletter draft issue contract and gated work list",
-      },
-    ],
-    spine: ["intent", "authority", "operation", "proof", "state"],
-    sections: [
-      {
-        kind: "opening",
-        body: "every coding agent i use eventually runs into the same boring problem: it can make a change, but it needs a system around it for responsibility.",
-        source_refs: ["AGENTS.md", "docs/admin-v2-architecture.md"],
-        thread_beat:
-          "the useful split is simple: safe lanes can move, hard gates come back to me.",
-      },
-      {
-        kind: "note",
-        heading: "safe lanes versus hard gates",
-        body: "i want agents to move quickly when the blast radius is small. if the work is safe, the agent should inspect, edit, verify, commit, push, and open a PR without waiting on me. if the work touches dns, auth, secrets, env, production content, outbound sends, live endpoints, payments, or destructive cleanup, it needs explicit authority. the control plane makes that split obvious before anything gets touched.",
-        source_refs: ["AGENTS.md"],
-        thread_beat:
-          "safe lanes let agents move without asking me for permission every five minutes.",
-      },
-      {
-        kind: "receipt",
-        label: "proof beats trust",
-        body: "if a decision only lives in chat, it barely exists. i want the useful state in commits, PRs, handoff docs, deploy runs, screenshots, and blocked-action lists. the next agent should be able to continue from the artifact instead of guessing what happened in the previous conversation.",
-        source_refs: ["AGENTS.md", "docs/content-admin-editor-brief.md"],
-        thread_beat:
-          "proof is what lets the next agent continue without inheriting chat memory.",
-      },
-      {
-        kind: "note",
-        heading: "NEEDS-ANI as a human syscall queue",
-        body: "the human queue should be tiny. i do not want a dashboard full of agent homework. i want the few calls only i can make: approve this, choose between these options, do this account-side action, provide this secret, decide whether this public copy sounds like me.",
-        source_refs: ["AGENTS.md", "docs/admin-v2-architecture.md"],
-        thread_beat:
-          "a good human queue is small enough to answer from my phone.",
-      },
-      {
-        kind: "worklog",
-        heading: "admin as the control plane",
-        body: "this is why admin.anipotts.com starts read-only. i need one place where content, fleet state, handoffs, approvals, and proof are legible. only after that is boring do i want write buttons. live controls come after the foundation.",
-        items: [
-          "read-only operator dashboard first",
-          "content inventory and preview states next",
-          "guarded writes only after the authority model is real",
-          "live gates visible before live controls exist",
-        ],
-        source_refs: [
-          "docs/admin-v2-architecture.md",
-          "docs/content-admin-editor-brief.md",
-          "apps/admin/src/data/content.ts",
-        ],
-        thread_beat:
-          "the control plane maps what i meant to what was allowed to what changed.",
-      },
-    ],
-    claims: [
-      {
-        claim:
-          "agents in this repo can complete safe lanes end to end but must stop at true gates",
-        status: "source_backed",
-        source_refs: ["AGENTS.md"],
-      },
-      {
-        claim: "NEEDS-ANI is defined as a narrow human syscall queue",
-        status: "source_backed",
-        source_refs: ["AGENTS.md"],
-      },
-      {
-        claim:
-          "admin content work should move from read-only inventory to previews before publish paths",
-        status: "source_backed",
-        source_refs: ["AGENTS.md", "docs/content-admin-editor-brief.md"],
-      },
-      {
-        claim:
-          "admin.anipotts.com should become the practical web surface for editing public-site text, previewing changes, reviewing fleet state, and returning decisions",
-        status: "source_backed",
-        source_refs: ["AGENTS.md"],
-      },
-    ],
-    blocked_actions: [
-      "write production d1 issue",
-      "publish archive page",
-      "send test email",
-      "broadcast to subscribers",
-      "change DNS",
-      "change auth or Cloudflare Access",
-      "modify env or secrets",
-      "change workers or endpoints",
-      "add admin write path",
-    ],
-    preview_notes:
-      "newsletter-first outline with thread-compatible beats. keep practical, source-backed, and specific to anipotts-com/admin.",
-  },
-];
+export {
+  newsletterDrafts,
+  newsletterDraftsSource,
+} from "@anipotts/content/admin";
