+++ docs/newsletter-content-structure.md
@@ -18,7 +18,8 @@ admin writes, or production d1 changes.
 | issue table         | `drizzle/migrations/0005_newsletter_system.sql:newsletter_issues`               | future storage shape for issue rows                                       | schema reference only               |
 | system brief        | `docs/newsletter-system.md`                                                     | infrastructure, resend, compliance, and rollout gates                     | authoritative for live-path gates   |
 | content inventory   | `docs/content-admin-editor-brief.md`                                            | public-site editable-content inventory and newsletter backfill candidates | admin planning source               |
-| read-only admin     | `apps/admin/src/data/content.ts`                                                | static inventory and preview rows for current content surfaces            | preview model only                  |
+| shared content      | `packages/content/src/admin/*`                                                  | static inventory, preview rows, and newsletter issue draft data           | preview model only                  |
+| read-only admin     | `apps/admin/src/data/content.ts`                                                | compatibility re-export for current admin routes                          | preview model only                  |
 | newsletter admin    | `apps/admin/src/data/newsletter.ts` and `apps/admin/src/pages/newsletter.astro` | static issue draft review inside canonical Astro admin                    | preview model only                  |
 | newsletter worker   | `workers/newsletter/*`                                                          | async send worker and queue consumer                                      | out of scope for this lane          |
 
@@ -83,7 +84,7 @@ do not load it into d1 or an admin write endpoint without separate authority.
     "AGENTS.md",
     "docs/content-admin-editor-brief.md",
     "docs/newsletter-system.md",
-    "apps/admin/src/data/content.ts"
+    "packages/content/src/admin/content.ts"
   ],
   "claims": [
     {
